+++ dffml/source/db.py
@@ -3,7 +3,6 @@ from typing import Type, AsyncIterator, List
 
 from dffml.base import config, BaseConfig
 from dffml.db.base import BaseDatabase, Condition
-from dffml.db.sqlite import SqliteDatabase
 from dffml.record import Record
 from dffml.source.source import BaseSource, BaseSourceContext
 from dffml.util.entrypoint import entrypoint
@@ -11,8 +10,7 @@ from dffml.util.entrypoint import entrypoint
 
 @config
 class DbSourceConfig(BaseConfig):
-    db_implementation: str
-    db_config: any
+    db: BaseDatabase
     table_name: str
     model_columns: List[str]
 
@@ -43,24 +41,14 @@ class DbSourceContext(BaseSourceContext):
                     key_value_pairs[key] = 1
             else:
                 key_value_pairs[key] = record.data.__dict__[key]
-        async with self.parent.db_impl() as db_ctx:
-            # Insert AND update since some tests don't insert before updating
-            # TODO: Maybe change the sqlite update query to use INSERT INTO ... ON DUPLICATE KEY UPDATE
-            # Note that this would impose all other db implementations to use "update or create if not exists" functionality
-            await db_ctx.insert(self.parent.config.table_name, key_value_pairs)
-            await db_ctx.update(
-                self.parent.config.table_name,
-                key_value_pairs,
-                [[Condition("key", "=", record.key)]],
+        async with self.parent.db() as db_ctx:
+            await db_ctx.insert_or_update(
+                self.parent.config.table_name, key_value_pairs
             )
-            results = [
-                row
-                async for row in db_ctx.lookup(self.parent.config.table_name)
-            ]
         self.logger.debug("update: %s", await self.record(record.key))
 
     async def records(self) -> AsyncIterator[Record]:
-        async with self.parent.db_impl() as db_ctx:
+        async with self.parent.db() as db_ctx:
             async for result in db_ctx.lookup(self.parent.config.table_name):
                 yield self.convert_to_record(result)
 
@@ -86,7 +74,7 @@ class DbSourceContext(BaseSourceContext):
 
     async def record(self, key: str):
         record = Record(key)
-        async with self.parent.db_impl() as db_ctx:
+        async with self.parent.db() as db_ctx:
             try:
                 row = await db_ctx.lookup(
                     self.parent.config.table_name,
@@ -122,22 +110,13 @@ class DbSourceContext(BaseSourceContext):
 class DbSource(BaseSource):
     CONFIG = DbSourceConfig
     CONTEXT = DbSourceContext
-    db_impl: BaseDatabase
 
     def __init__(self, cfg: Type[BaseConfig]) -> None:
         super().__init__(cfg)
 
     async def __aenter__(self) -> "DbSource":
-        # Instantiate a subclass of BaseDatabase using "db_implementation" from DbSourceConfig
-        self.db_impl = None
-        if self.config.db_implementation == "sqlite":
-            self.db_impl = SqliteDatabase(self.config.db_config)
-        # TODO: Add more implementations. Is there a better way to do this?
-        else:
-            raise RuntimeError(
-                f"No implementation named {self.config.db_implementation} found for DbSource abstraction"
-            )
+        self.db = await self.config.db.__aenter__()
         return self
 
     async def __aexit__(self, exc_type, exc_value, traceback):
-        await self.db_impl.__aexit__(exc_type, exc_value, traceback)
+        await self.db.__aexit__(exc_type, exc_value, traceback)